wip move readonly stuff into shared module
